Refactor name conventions
