--- client/src/services/user.service.js
@@ -1,20 +1,0 @@
-import Axios from 'axios';
-import authHeader from './authHeader';
-
-const BASE_URL = 'http://localhost:3001/api/';
-
-class UserService {
-	async getAnswers() {
-		const { data: { answers } } =
-			await Axios.get(BASE_URL + 'user/answers', { headers: authHeader() });
-		return answers || [];
-	}
-
-	async answer(selectedAnswer) {
-		return await Axios.post(BASE_URL + 'user/answer', {
-			selectedAnswer
-		}, { headers: authHeader() });
-	}
-}
-
-export default new UserService();
