+++ client/src/views/contents/login/index.jsx
@@ -3,7 +3,7 @@ import { useHistory } from 'react-router-dom';
 import Popup from '../../../components/utils/popup';
 import TextField from '../../../components/utils/textField';
 import { AuthContext } from '../../../contexts/AuthContext';
-import AuthService from '../../../services/auth.service';
+import * as AuthService from '../../../services/auth';
 import { Container, FormContainer, LoginButton, StyledLink } from './styles';
 import { Typography } from '@material-ui/core';
